Tidy go mod
