+++ faucet/main.go
@@ -6,8 +6,10 @@ import (
 	"flag"
 	"fmt"
 	"os"
+	"os/signal"
 	"regexp"
 	"strconv"
+	"syscall"
 
 	"github.com/gnolang/faucet"
 	tm2Client "github.com/gnolang/faucet/client/http"
@@ -16,6 +18,7 @@ import (
 	"github.com/gnolang/gno/tm2/pkg/std"
 	"github.com/pelletier/go-toml"
 	"go.uber.org/zap"
+	"golang.org/x/sync/errgroup"
 
 	"github.com/peterbourgon/ff/v3/ffcli"
 )
@@ -152,4 +155,100 @@ func readConfig(path string) (*config.Config, error) {
 	}
 
 	return &faucetConfig, nil
+}
+
+type cmdLogger struct {
+	logger *zap.Logger
+}
+
+func newCommandLogger(logger *zap.Logger) *cmdLogger {
+	return &cmdLogger{
+		logger: logger,
+	}
+}
+
+func (c *cmdLogger) Info(msg string, args ...interface{}) {
+	if len(args) == 0 {
+		c.logger.Info(msg)
+
+		return
+	}
+
+	c.logger.Info(msg, zap.Any("args", args))
+}
+
+func (c *cmdLogger) Debug(msg string, args ...interface{}) {
+	if len(args) == 0 {
+		c.logger.Debug(msg)
+
+		return
+	}
+
+	c.logger.Debug(msg, zap.Any("args", args))
+}
+
+func (c *cmdLogger) Error(msg string, args ...interface{}) {
+	if len(args) == 0 {
+		c.logger.Error(msg)
+
+		return
+	}
+
+	c.logger.Error(msg, zap.Any("args", args))
+}
+
+type waitFunc func(ctx context.Context) error
+
+// waiter is a concept used for waiting on running services
+type waiter struct {
+	ctx    context.Context
+	cancel context.CancelFunc
+
+	waitFns []waitFunc
+}
+
+// newWaiter creates a new waiter instance
+func newWaiter() *waiter {
+	w := &waiter{
+		waitFns: []waitFunc{},
+	}
+
+	w.ctx, w.cancel = signal.NotifyContext(
+		context.Background(),
+		os.Interrupt,
+		syscall.SIGINT,
+		syscall.SIGTERM,
+		syscall.SIGQUIT,
+	)
+
+	return w
+}
+
+// add adds a new wait service
+func (w *waiter) add(fns ...waitFunc) {
+	w.waitFns = append(w.waitFns, fns...)
+}
+
+// wait blocks until all added wait services finish
+func (w *waiter) wait() error {
+	g, ctx := errgroup.WithContext(w.ctx)
+
+	g.Go(func() error {
+		<-ctx.Done()
+		w.cancel()
+
+		return nil
+	})
+
+	for _, fn := range w.waitFns {
+		fn := fn
+
+		g.Go(
+			func() error {
+				return fn(ctx)
+			},
+		)
+	}
+
+	return g.Wait()
 }
